headers and misc

+++ src/xrGame/ui/UIInventoryUtilities.cpp
@@ -346,7 +346,7 @@ LPCSTR InventoryUtilities::GetTimePeriodAsString(LPSTR _buff, u32 buff_sz, ALife
     int cnt = 0;
     _buff[0] = 0;
 
-    u8 yrdiff = ((year2 - year1) * 12);
+    u32 yrdiff = ((year2 - year1) * 12);
     if (month1 != month2 || yrdiff > 0)
         cnt = xr_sprintf(_buff + cnt, buff_sz - cnt, "%d %s", month2 + (yrdiff - month1), *CStringTable().translate("ui_st_months"));
